Add FlaggedReview interface and components for displaying flagged reviews
